pyproject.toml
[tool.poetry]
name = "vsporte"
version = "0.1.0"
description = ""
authors = ["Mat0mba24 <mat0mba@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.10.1"
sqlalchemy = "^2.0.36"
fastapi = "^0.115.4"
asyncpg = "0.27.0"
pydantic-settings = "^2.6.1"
python-jose = "^3.3.0"
uvicorn = "^0.32.0"
passlib = "^1.7.4"
betterlogging = "^1.0.0"
bcrypt = "4.0.1"
python-json-logger = "^2.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
